re-org tests, clean-up
