+++ src/bbs2gh/Program.cs
@@ -29,12 +29,12 @@ namespace OctoshiftCLI.BbsToGithub
                 .AddSingleton<EnvironmentVariableProvider>()
                 .AddSingleton<BbsApiFactory>()
                 .AddSingleton<GithubApiFactory>()
-                .AddSingleton<HttpDownloadServiceFactory>()
                 .AddSingleton<RetryPolicy>()
                 .AddSingleton<IAzureApiFactory, AzureApiFactory>()
                 .AddSingleton<IBlobServiceClientFactory, BlobServiceClientFactory>()
                 .AddSingleton<AwsApiFactory>()
                 .AddSingleton<VersionChecker>()
+                .AddSingleton<HttpDownloadServiceFactory>()
                 .AddSingleton<FileSystemProvider>()
                 .AddSingleton<DateTimeProvider>()
                 .AddSingleton<IVersionProvider, VersionChecker>(sp => sp.GetRequiredService<VersionChecker>())
